revert moving of config string reading
