+++ pkg/service/roomallocator_test.go
@@ -17,7 +17,7 @@ import (
 
 func TestCreateRoom(t *testing.T) {
 	t.Run("ensure default room settings are applied", func(t *testing.T) {
-		conf, err := config.NewConfig(nil, nil, true)
+		conf, err := config.NewConfig("", true, nil, nil)
 		require.NoError(t, err)
 
 		node, err := routing.NewLocalNode(conf)
@@ -32,7 +32,7 @@ func TestCreateRoom(t *testing.T) {
 	})
 
 	t.Run("reject new participants when track limit has been reached", func(t *testing.T) {
-		conf, err := config.NewConfig(nil, nil, true)
+		conf, err := config.NewConfig("", true, nil, nil)
 		require.NoError(t, err)
 		conf.Limit.NumTracks = 10
 
@@ -48,7 +48,7 @@ func TestCreateRoom(t *testing.T) {
 	})
 
 	t.Run("reject new participants when bandwidth limit has been reached", func(t *testing.T) {
-		conf, err := config.NewConfig(nil, nil, true)
+		conf, err := config.NewConfig("", true, nil, nil)
 		require.NoError(t, err)
 		conf.Limit.BytesPerSec = 100
